requirements.txt
torch
pillow
numpy
opencv-python
scipy
matplotlib
blendmodes